Refactor project structure for future development
